MiniMax function now works!
U has to make the AI go first in order for miniMax to work but we have an unbeatable AI now! We also have a blinking animation for illegal moves. Lastly, I reduced the difficulty on the medium AI to only block the player 75% of the time so the player can have a winning chance. This project is now complete. Might clean up some code!